requirements.txt
flask>=3.0.0
python-docx>=1.1.0
markdown-it-py>=3.0.0
mdit-py-plugins>=0.4.0
pywebview>=5.0
